rhp4: combine client and server packages
