+++ rhp/v4/options.go
@@ -1,4 +1,4 @@
-package host
+package rhp
 
 import (
 	"time"
@@ -6,18 +6,18 @@ import (
 	"go.uber.org/zap"
 )
 
-// A ServerOption sets an option on a Server.
-type ServerOption func(*Server)
+// An Option sets an option on a Server.
+type Option func(*Server)
 
 // WithLog sets the logger for the server.
-func WithLog(log *zap.Logger) ServerOption {
+func WithLog(log *zap.Logger) Option {
 	return func(s *Server) {
 		s.log = log
 	}
 }
 
 // WithPriceTableValidity sets the duration for which a price table is valid.
-func WithPriceTableValidity(validity time.Duration) ServerOption {
+func WithPriceTableValidity(validity time.Duration) Option {
 	return func(s *Server) {
 		s.priceTableValidity = validity
 	}
@@ -25,7 +25,7 @@ func WithPriceTableValidity(validity time.Duration) ServerOption {
 
 // WithContractProofWindowBuffer sets the buffer for revising a contract before
 // its proof window starts.
-func WithContractProofWindowBuffer(buffer uint64) ServerOption {
+func WithContractProofWindowBuffer(buffer uint64) Option {
 	return func(s *Server) {
 		s.contractProofWindowBuffer = buffer
 	}
